如果需要改代码自用请回滚到1.41X封版使用 1. 增加回调方式 2. 优化所有新增文件的后缀格式、文件名

+++ progresslibrary/src/main/java/com/zhongjh/progresslibrary/api/MaskProgressApi.kt
@@ -68,7 +68,7 @@ interface MaskProgressApi {
      * @param filePath 音频文件地址
      * @param length   音频文件长度
      */
-    fun addAudioStartUpload(filePath: String, length: Int)
+    fun addAudioStartUpload(filePath: String, length: Long)
 
     /**
      * 添加音频网址数据
